% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-annotate_doco_with_gene.R
\name{blessy.annotate_doco_with_gene}
\alias{blessy.annotate_doco_with_gene}
\title{Annotate the DoCo groups with gene ID associated with each transcript}
\usage{
blessy.annotate_doco_with_gene(count_se, doco_se)
}
\arguments{
\item{count_se}{A SummarizedExperiment object containing transcript counts with gene IDs.}

\item{doco_se}{A SummarizedExperiment object containing DoCo information based on transcript IDs.}
}
\value{
A SummarizedExperiment object with transcript count with DoCo fully annotated (with gene IDs)
}
\description{
Annotate the DoCo groups with gene ID associated with each transcript
}
\examples{
# Example usage:
# final_se <- annotate_tx_with_doco(count_se, doco_se)
}
