% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-annotate_gene.R
\name{blessy.annotate_gene}
\alias{blessy.annotate_gene}
\title{Annotate transcripts with Gene information using the GTF file}
\usage{
blessy.annotate_gene(doco_se, gtf_file_path)
}
\arguments{
\item{doco_se}{A SummarizedExperiment object containing DoCo (domain combinations) for each transcript.}

\item{gtf_file_path}{A string specifying the path to the gene annotation GTF file.}
}
\value{
A SummarizedExperiment object containing the transcript-to-gene mapping with updated DoCo information.
A CSV file "dict.csv" is saved in the \code{./out} directory.
}
\description{
This function maps transcripts in the DoCo SummarizedExperiment to genes using a GTF annotation file,
adds GeneID information to the domain combination (DoCo), and saves the result as a CSV file.
}
\examples{
dict_se <- blessy.annotate_gene(doco_se, "../../data/Homo_sapiens.GRCh38.111.gtf")

}
