% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-annotate_tx_with_doco.R
\name{blessy.annotate_tx_with_doco}
\alias{blessy.annotate_tx_with_doco}
\title{Annotate transcripts with domain combinations (DoCo) using the mapping GRanges object}
\usage{
blessy.annotate_tx_with_doco(gr)
}
\arguments{
\item{gr}{A GRanges object containing transcript IDs and domains with their genomic coordinates.}
}
\value{
A SummarizedExperiment object containing the DoCo for each transcript.
}
\description{
Annotate transcripts with domain combinations (DoCo) using the mapping GRanges object
}
\examples{
#doco_se <- blessy.annotate_tx_with_doco(gr)
}
