% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-generate_dict.R
\name{blessy.generate_final_dict}
\alias{blessy.generate_final_dict}
\title{Generate a Final Dictionary}
\usage{
blessy.generate_final_dict(final_se)
}
\arguments{
\item{final_se}{A SummarizedExperiment object that contains transcript counts and DoCo information.}
}
\value{
A data frame with three columns: Gene, Transcript, and DoCo.
}
\description{
This function creates a dictionary data frame containing gene, transcript, and DoCo information from a SummarizedExperiment object.
}
\examples{
# Assuming `final_se` is your SummarizedExperiment object
final_dict <- blessy.generate_final_dict(final_se)
}
