% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-get_transcript_track.R
\name{blessy.get_transcript_track}
\alias{blessy.get_transcript_track}
\title{Get Transcript Track Data and Save as BED File}
\usage{
blessy.get_transcript_track(genome, track)
}
\arguments{
\item{genome}{A string specifying the genome version (e.g., "hg38").}

\item{track}{A string specifying the track name (e.g., "wgEncodeGencodeBasicV44").}
}
\value{
A BED file saved in the \code{./out} directory with the transcript track data. The function prints a message confirming the save.
}
\description{
This function fetches transcript track data from the UCSC Genome Browser API and saves it as a BED file.
}
\examples{
# Example usage:
blessy.get_transcript_track("hg38", "wgEncodeGencodeBasicV44")

}
