% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-load_mapping_txt.R
\name{blessy.load_transcript_domains}
\alias{blessy.load_transcript_domains}
\title{Load Mapping text file into a GRanges Object}
\usage{
blessy.load_transcript_domains(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the mapping file.}
}
\value{
A GRanges object containing the genomic ranges of the protein domains with transcript IDs as metadata.
}
\description{
This function reads a text file containing transcript IDs and protein domain information with genomic coordinates,
and loads the data into a `GRanges` object from the `GenomicRanges` package using regex.
}
\examples{
# Example usage:
# Load and test the GRanges object
gr <- load_transcript_domains("./data/mapping.txt")
}
