% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite.R
\name{blessy.output}
\alias{blessy.output}
\title{Write DoCo Count and Final Dictionary to CSV Files}
\usage{
blessy.output(results, doco_count_file, final_dict_file)
}
\arguments{
\item{results}{A list containing the `doco_count` matrix and `final_dict` data frame, as generated by the Blessy function.}

\item{doco_count_file}{The file path where the DoCo count matrix will be saved as a CSV file.}

\item{final_dict_file}{The file path where the final dictionary will be saved as a CSV file.}
}
\description{
This function takes the results from the Blessy function (a list containing the `doco_count` and `final_dict`)
and writes them to two separate CSV files.
}
\examples{
# Example usage:
results <- blessy("path/to/transcript_counts.txt", "path/to/domain_mapping.txt")
blessy.write_results_to_csv(results, "doco_count.csv", "final_dict.csv")
}
