% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-generate_doco_count.R
\name{blessy.generate_doco_count}
\alias{blessy.generate_doco_count}
\title{Generate the DoCo Count Matrix}
\usage{
blessy.generate_doco_count(final_se)
}
\arguments{
\item{final_se}{A SummarizedExperiment object that contains transcript counts and DoCo information.}
}
\value{
A data frame representing the DoCo count matrix, where each row corresponds to a unique DoCo with its aggregated read counts.
}
\description{
This function aggregates the read counts of transcripts belonging to the same DoCo from a SummarizedExperiment object.
}
\examples{
# Assuming `final_se` is your SummarizedExperiment object
#doco_count_matrix <- blessy.generate_doco_count(final_se)
head(doco_count_matrix)
}
