% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite.R
\name{blessy.output}
\alias{blessy.output}
\title{Write DoCo Count and Final Dictionary to CSV Files}
\usage{
blessy.output(results, doco_count_file, final_dict_file)
}
\arguments{
\item{results}{A list containing the \code{doco_count} matrix and \code{final_dict} data frame, as generated by the Blessy function.}

\item{doco_count_file}{The file name for the DoCo count matrix CSV (this will be saved in the \verb{/out} folder).}

\item{final_dict_file}{The file name for the final dictionary CSV (this will be saved in the \verb{/out} folder).}
}
\value{
None. The function writes the \code{doco_count} and \code{final_dict} to CSV files in the \verb{/out} directory.
}
\description{
This function takes the results from the Blessy function (a list containing the \code{doco_count} matrix and \code{final_dict} data frame)
and writes them to two separate CSV files in the \verb{/out} folder.
}
\details{
The function ensures that the \verb{/out} directory exists, and if not, it creates the directory.
The two CSV files, \code{doco_count} and \code{final_dict}, are then written to this folder.
}
\examples{
# Example usage:
results <- blessy("path/to/transcript_counts.txt", "path/to/domain_mapping.txt")
blessy.output(results, "doco_count.csv", "final_dict.csv")
}
